Fix: Enable EVPN only on IPv4 or RFC 8950 BGP sessions (#2813)
netlab currently does not support EVPN with IPv6 VTEPs, so we have to
limit the activation of EVPN AFs only to BGP sessions that also carry
IPv4 next hops.

The EVPN module has been modified to set 'evpn' flag on a BGP neighbor
only when 'ipv4' is defined on that neighbor (to be changed to a more
comprehensive test when we support VXLAN-over-IPv6). The flag is thus
set for real IPv4 BGP sessions, unnumbered sessions (because IPv4 is
set to True) and RFC 8950 sessions over IPv6 GUA addresses (IPv4 is
also set to True in that case).

Several device configuration templates have been modified as needed
to select the correct transport session to use for the EVPN AF.

Finally, the 'evpn-vxlan' transformation tests adds an IPv6-only
network to verify that the EBGP session running over it is not
carrying EVPN AF.

In a future patch, we migth change the 'evpn' attribute to 'ipv4',
'ipv6', or 'local_if' to make it easier for templates to select
the correct neighbor attribute.

Closes #2810