Transform community lists into type/value format (#2851)
This is a preparatory step for the introduction of extended/large
community lists. The final format of the community lists is
unchanged, but the transformation into 'type'/'value' format is
done earlier in the process as we'll use 'type'/'value' format
as the canonical format for the more complex community lists.

The final data structure is unchanged, which means that this
refactoring does not impact the device configuration templates
or any user-side functionality depending on the transformed
data structures.

The 'rp-clist-expansion' transformation test was modified to
include the single-entry box format and the canonical
type/value format to test all possible lab topology formats
that can be used to specify a standard clist.