Refactor the containerlab configuration template
The containerlab configuration template used a list of "extra" clab
attributes that would be copied into the template.

The new template reverses that logic; all clab attributes apart from
the "special" attributes defined in node_config_special are copied
verbatim, the special attributes are handled by the template.

This allows the users to add clab attributes that get copied into
the template without changing the template or replacing the
"node_config_attributes" list.