Fix: Copy relevant link module data to interfaces attached to a bridge
The 'convert multi access link into a hub-and-spoke with a bridge' code
copied the link data into bridge VLAN data. However, the VLAN code
removed the module data not present on the bridge node (i.e. most of
them) from the node VLAN, so they were never propagated to the
interfaces attached to the bridge node.

This fix copies relevant (present on node, copying allowed) module
data into interfaces attached to a bridge node when transforming the
multi-access link into the hub-and-spoke links.