Run host- or container configuration scripts in 'netlab initial'
This commit adds the ability to run container bash scripts either in
the container or in the host namespace. It's currently used to
configure Bird, dnsmasq, and Linux.

The core change is an extra step in the 'netlab initial' process
that calls provider-specific 'deploy_node_config' hook. That clab hook
executes a bash script (identified by its suffix) either on the host
(in container namespace) or in the container.

Other changes:
* netlab initial configuration deployment has been refactored
  into a separate module (deploy.py)
* Refactor the 'do we need Ansible' code into a set of generic
  functions that (a) identify the configs a user wants to deploy
  with 'netlab initial', (b) figures out whether all configs are
  covered with config scripts and (c) does that for the requested
  nodeset
* Use the refactored functions in configuration creation ('-o')
  and deployment, resulting in cleaner configuration creation

Minor fixes:
* Bird implementation no longer needs extra 'config.sh' template
  and 'clab.exec' parameter
* Bird loopback LLA configuration does not crash on second run
* Linux routing configuration no longer prints the final routing
  table
* netlab initial '-v' flag is used to set log.VERBOSE variable