Run host- or container configuration scripts in 'netlab initial' (#2903)
This commit adds the ability to run container bash scripts either in
the container or in the host namespace. It's currently used to
configure Bird, dnsmasq, and Linux.

The core change is an extra step in the 'netlab initial' process
that calls provider-specific 'deploy_node_config' hook. That clab hook
executes a bash script either on the host (in container namespace)
or in the container.

The type of script is identified by the 'mode' suffix (sh for
container-side scripts, ns for host-side namespace scripts)
in config_template values. The config_template value can have
empty "mapped_file" component for host-side scripts

Other changes:
* netlab initial configuration deployment has been refactored
  into a separate module (deploy.py)
* Refactor the 'do we need Ansible' code into a set of generic
  functions that (a) identify the configs a user wants to deploy
  with 'netlab initial', (b) figures out whether all configs are
  covered with config scripts and (c) does that for the requested
  nodeset
* Use the refactored functions in configuration creation ('-o')
  and deployment, resulting in cleaner configuration creation
* The 'create_extra_file_mappings' function has been renamed to
  'find_node_templates' because that's all it does now
* The configs-to-binds step has been removed from the
  'find_node_templates' function because it's clab-specific
* The 'create_extra_files' function has been renamed to
  'create_node_files' and no longer uses clab.binds
* An extra step has been added to ensure clab.binds refer to
  valid local files
* We had to swap device quirks and provider post-transform hook

A device quirk (Arista cEOS) might change provider 'config_templates'
dictionary, which means that the device quirks must be called before
the 'config_templates' are compiled into binds.

Minor fixes:
* Bird implementation no longer needs extra 'config.sh' template
  and 'clab.exec' parameter.
* Bird loopback LLA configuration does not crash on second run
* Linux routing configuration no longer prints the final routing
  table
* netlab initial '-v' flag is used to set log.VERBOSE variable
* Remove 'is topology valid' checks from configuration deployment
  (it's always valid)
* Rework the final status/message 'netlab initial' printouts to ensure
  there are always enough empty lines