Disable DHCP server on tested devices in DHCP client integration tests
When testing the 'dnsmasq' device as a DHCP client, the DHCP client
flag is removed from the interface (because the device is by default a
DHCP server), resulting in validation failure.

The plugin introduced in this commit removes the 'dhcp.server' flag
from nodes in the 'dhcp_clients' group to ensure the DHCP clients are
never also DHCP servers, causing the DHCP feature check to fail on
dnsmasq DHCP clients.