Recreate node config files in netlab initial if templates changed
This commit resolves #2905 and adds automatic check-and-recreate
functionality to 'netlab initial'. Before creating configuration
files or deploying them, the templates are searched again (in case
the user added/removed a custom template) and the output files are
recreated if the underlying templates changed.

This functionality is controlled by the '--no-refresh' flag to allow
'netlab up' to run 'netlab initial' without checking files that were
created milliseconds ago.

Other changes:

* Template functions were moved from netsim.cli.initial.config to
  netsim.cli.initial.template
* 'netlab config' uses _template_cache instead of config_templates
  to create configuration files