Use 'unprovisioned' group to skip already-deployed nodes
'netlab initial' can use provider-specific deployment mechanisms for
individual nodes (example: Linux hosts, bird). These nodes do not have
to be configured by Ansible, and we solved that with a variable
telling Ansible to skip particular modules.

This commit optimizes that process: a node that does not need Ansible
deployment is excluded from Ansible inventory -- it's placed in a
subgroup of the 'unprovisioned' group (a subgroup is used to make
it easier to manipulate the inventory).